Remove the application registry.
The application registry (a mapping from text labels to `Application`
classes) was initially conceived as a way to extend `gsub`-like
commands: it should have provided a way to map a "user-friendly"
application name to a class.  Now that `gsub` is gone and we're taking
a different approach, this code is no longer needed.

Remove, for sake of simplification.

  modified:   ../../gc3libs/application/__init__.py
  modified:   ../../gc3libs/application/crypto.py
  modified:   ../../gc3libs/application/demo.py
  modified:   ../../gc3libs/application/gamess.py
  modified:   ../../gc3libs/application/rosetta.py
  modified:   ../../gc3libs/application/turbomole.py

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@1921 d0f2d5a0-84c0-18d2-3862-8add1af26540