New configuration option `private_credentials_copy` for `[auth/*]` sections.
If the new option is in effect, `GridAuth` instances will make a copy
of the SLCS cert and VOMS proxy in a directory (specified with option
`private_copy_directory`), encrypting the X.509 key with a random
password.

This provides a more secure alternative to `remember_password=yes`,
as the password that we keep in memory is a random one and not the
main SWITCHaai password.

This commit also introduces a way for Core/Engine objects to set
additional initialization parameters on the `Auth` objects, so we can
use the session directory as a credentials store, and we do not have
to fix the private credentials store in the configuration file.

  modified:   gc3libs/authentication/__init__.py
  modified:   gc3libs/authentication/grid.py
  modified:   gc3libs/cmdline.py

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@2030 d0f2d5a0-84c0-18d2-3862-8add1af26540