Easier-to-read syntax for chained access methods.
The new syntaxes `GET.item(...)` and `GET.attr(...)` take the place of
`value_at_item()` and `value_of()`.  The `GET` versions can be chained
left-to-right, which more naturally reflects the writing way of Python
expressions: `GET.attr('inputs').item(0)` is the accessor for Python
attribute `.inputs[0]`.

More magic should actually be possible, so to shorten the accessor
expression to `GET.inputs[0]`.

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@2362 d0f2d5a0-84c0-18d2-3862-8add1af26540