Fix "bug" causing multiple `nosetests` clones to be run while testing ShellcmdLrms.
The bug is cause by the fact that in order to avoid creation of zombie
processes we have to *daemonize*, which basically means that we have
to do something like:

if not posix.fork()
    # the child
    if not posix.fork():
        # the nephew
        os.execlp(<the program we want to run>)
    else: # still the child
        # exits, so that the nephew will not become a zombie
        sys.exit(0)

However, sys.exit() basically raises a SystemExit exception, which is
catched (and then "ignored") by nose, so we would end up with a lot of
clones of the nosetests programs running at the same time.

To avoid this we call os.execlp('/bin/true') instead, which will
overwrite the current instance of nosetests with /bin/true, which will
exit without problem.

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@2388 d0f2d5a0-84c0-18d2-3862-8add1af26540