New `gsession` command to manage sessions
This will address issue 210: http://code.google.com/p/gc3pie/issues/detail?id=210&can=3
Moreover, now Session's ctor accept a `create` parameter in order to avoid creation of invalid/mispelled sessions when called Session() with a non-existent session name. This is particolarly useful for gsession which must be able to check if the session exists or not.

	modified:   gc3libs/session.py
	modified:   gc3utils/commands.py
	modified:   setup.py

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@2407 d0f2d5a0-84c0-18d2-3862-8add1af26540