Another variation on Fliege's algorithm.
Remove the "perturbation" variant of Fliege's algorithm introduced in last commit:
it only works with triangular matrices.

However, we now allow for the vector population to be arbitrary; in
particular it can be larger than the matrix size (up to n*(n+1)/2),
which seems to improve the reliability of the procedure.

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@3323 d0f2d5a0-84c0-18d2-3862-8add1af26540