Update issue 372 Status: Fixed
Change the log severity from `ERROR` to `INFO` in `_load_session()`
when the `store.url` filename is not found in the directory, since at
this point we can't know yet if this is an error or not.

When a new `Session` object is created and the directory exists but it
does not contain all the required files, an exception is raised from
`_load_session()`, which is catched by the constructor that will
decide if the session can still be created or not, and raise a proper
exception if needed.

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@3360 d0f2d5a0-84c0-18d2-3862-8add1af26540