Remove loading of system's known hosts file. Note that this introduce a possible security issue, since we do not compare the host key with the database of known hosts.
However, working with VMs running on a cloud, especially a private
cloud like Hobbes, will cause quite often conflicts with new VM with
new ssh host keys getting an IP we already connected to in the past,
and thus causing a connection error.

I couldn't find a better way to deal with this issue, a warning would
be better, but for now it's better to silently accept any key rather
that NOT having gc3pie working with a cloud.

Fix issue #389.

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@3451 d0f2d5a0-84c0-18d2-3862-8add1af26540