Add `resourcedir` option to `ShellCmdLrms` constructor. This is used to modify the default value for the directory containing job informations.
This option is mainly needed for running tests: having multiple
ShellCmdLrms on the same machine sharing the same directory will work,
but it will cause each resource being able to see the jobs submitted
by the other resources. This is a problem while running multiple `tox`
in parallel on the same machine, since tests will make some
assumptions on the available resources.

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@3487 d0f2d5a0-84c0-18d2-3862-8add1af26540