Move handling of `LRMSSkipSubmissionToNextIteration` into the scheduler code.
It cannot be handled in `Engine`: the scheduler needs to decide whether
to retry the same task on a different resource or just keep going on.

The `first_come_first_serve` scheduler ignores the
`LRMSSkipSubmissionToNextIteration` condition (so it will retry on the
same resource when it's called some time in the future), just like the
`Engine` code did before the "new scheduler" merge.

git-svn-id: http://gc3pie.googlecode.com/svn/trunk@3740 d0f2d5a0-84c0-18d2-3862-8add1af26540